Collate Rt estimates
Adds script to collate Rt estimates from different data sources within one region, creating a new folder (`collate/`) with one csv output.

Uses a new dependency, `purrr` - but open to a re-write to avoid this if necessary.

In general obviously happy for this to be edited to better fit with the infrastructure.